Split parser_visitor.py
Split parser_visitor.py into more managable files
